plugins: gazebo_motor_failure_model
This plugin adds motor failure feature to the existing motor_model. Use this plugin in your model to include the motor model and also the motor failure feature. This plugin subscribes to rostopic `/gazebo/motor_failure/motor_number` over a ROS node and takes the motor number (integer) of the motor to fail.