typhoon_h480: increase the gimbal pitch range
When the drone is moving forward it is pitched down, however, we still
want to have the camera/gimbal pointing level forwards. Therefore, we
need to allow the gimbal to be able to point higher than horizontal.
Similarly, we need to increase the range down for when the drone moves
backwards.