gazebo_imu_plugin: use queue of 10 for IMU msg pub
The problem that we're trying to work around here is that sometimes the
mavlink plugin misses an IMU sample especially if the CPU load is high.

What we can do is wait in `OnUpdate` in the mavlink plugin until we have
received the callback from the publication here. However, if we're
unlucky we won't even get to `OnUpdate` in the mavlink plugin before the
IMU plugin publishes yet another sample. In that case we overwrite the
sample if we don't have a queue > 1 which is what is changed in this
commit.