gazebo_mavlink_interface: add lockstep
This adds the option for lockstep simulation. Lockstep is off be default
unless `enable_lockstep` is set to true in the vehicle's SDF file.

Basically, lockstep means that PX4 is "stepped" forward by each
hil_sensor MAVLink message containing a timestamp which is then the time
set for PX4. Then, this plugin waits for the actuator_control message
coming back from PX4. Once that has arrived, we move on which means we
resume the simulation to get a new IMU sample.

Therefore there are two locations where we wait:
1. In `OnUpdate` we wait for the next IMU sample in case the callback
   for the IMU message subscription has not triggered yet.
2. Then later in `OnUpdate` we wait for the actuator control message to
   come back from PX4. Basically we use `poll` and `recv` until we have
   parsed the message that we're looking for.

Note that we "freewheel", so we send hil_sensor messages without waiting
for actuator_control messages for initialization until PX4 starts
sending actuator_control messages. Once we have received the first
control message, we start the "lockstepping".