standard_vtol: fix vibration with ground
By adapting mass and inertial moments to something closer to the actual
delta quad model, the vibration of the delta quad while landed stopped.

The inertial moments are based on the superposition of two  cuboids,
one for the wing and one for the body.

Wing:
    height: 0.1 m
    width:  0.2 m
    depth:  0.9 m
    mass:   4.0 kg

Body:
    height: 0.1 m
    width:  2.35 m
    depth:  0.9 m
    mass:   1.0 kg

In order to get the model to fly again in SITL I also needed to increase
the wing areas as well as the `rad_to_cl` control parameters which
represent the lift coefficient given the angle of attack, as well as
make the pusher (called puller?) motor faster.

At this point, the model is still far from reality but at least it
doesn't vibrate on the ground any longer. Presumably, the model settings
are actually from Fun Cub, so a standard VTOL, and not the actual delta
quad.