Set wind_pluggin to zero for Gazebo Plane model
As documented here https://github.com/PX4/sitl_gazebo/issues/287 the Gazebo wind_pluggin doesn't work properly for the Gazebo Plane model.  This PR sets the wind_pluggin values to zero for the Plane model so that it will fly correctly in simulation.  In the long term, we should figure out exactly why the wind_pluggin does not work properly with gazebo plane model.  I suspect it might have something to do with improper or incomplete modeling of the airplane aerodynamics as it relates to the wind_pluggin, but I don't have the coding skills to really figure this out for sure. 
Questions:
1) if this PR is viewed as too much a hack, feel free to reject it, or suggest improvements.
2) maybe there are other vehicle models that are also affected?  Should we fix it for them too?