Fix HITL by sending heartbeats
PX4 no longer starts the mavlink instance on USB by default. This means
that the simulator needs to send heartbeats to initialize it.

Sending the heartbeats at 1 Hz however is not enough as we might fill up
all the buffers in less than a second which means PX4 never sees a
heartbeat, and won't initialize. Therefore, we send heartbeats in every
iteration initially until we have received a heartbeat back.