Multiple aircrafts can now exist inside of Gazebo.
This problem originates from setting global real_time_update_rate without checking if it has been set before.

To fix this, we check if we've done the calculation before, then do the calculation to update, then then check to see if the updated real_time_update_rate  is still valid.