airspeed_plugin: fix sign of differential pressure
It didn't seem right that an airspeed sensor measures a positive
airspeed both facing against the wind and away from the wind.

My best guess is that we are missing the sign in the differential
pressure which I have added in this commit.

This fix has been found while investigating wind and airspeed for
Wingcopter (wingcopter.com).